Rename worker package to workers
